django_chat/requirements.txt
python-dotenv==0.15.0;
psycopg2-binary==2.9.5;
pyOpenSSL==23.0.0;
pillow==8.3.0;
django-crispy-forms==1.13.0;
channels==3.0.1;
channels-redis==3.3.1;
djangorestframework==3.13.0;